+++ src/Exceptions/EsiRateLimitedException.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Seatplus\EsiClient\Exceptions;
 
-class EsiRateLimitedException extends \RuntimeException
+class EsiRateLimitedException extends \RuntimeException implements EsiClientException
 {
     public function __construct(public readonly int $retryAfter = 60)
     {