+++ src/Exceptions/InvalidAuthenticationException.php
@@ -4,4 +4,4 @@ declare(strict_types=1);
 
 namespace Seatplus\EsiClient\Exceptions;
 
-class InvalidAuthenticationException extends \Exception {}
+class InvalidAuthenticationException extends \Exception implements EsiClientException {}